requirements.txt
futures==3.0.5
numpy==1.25.2
pandas==2.0.3
pygame==2.5.1
python-dateutil==2.8.2
python-mnist==0.7
pytz==2023.3
six==1.16.0
tzdata==2023.3
